비동기 처리  적용
