+++ config/mysql.md
@@ -1,0 +1,7 @@
+# mysql
+
+# mysql 과 mysql2 차이점
+
+## createPool 과 createConnection 차이
+
+https://velog.io/@beberiche/Javascript%EC%97%90%EC%84%9C-mySQL-%EC%8B%9C%EC%9E%91%ED%95%98%EA%B8%B0
